AutoTunerDecisionDescription - auto tuning actions are saved with their name, the databaseName, the time the action occurred and the reason the auto decision was made Document database holds a  fixed size queue of all the auto decisions made for the database.
Memory Statistics - holds a fixed size queue of all the low memory calls, whenever there is a call for a low memory handler we keep track of the time it started, duration and if the low memory handler was called automatically
or manually

All the HandleLowMemory method are now returning LowMemoryHandlerStatistics - in order to keep track of the Name of the process who call it, the name of the database and the a Summary of the low memory action.

CpuStatistics- holds a queue of all the CpuUsage changes Normal, Low or High. The queue holds the last state of the CPU usage only once- for example, if there was a long time of low CPU it won't be saved in the queue more than once, so the last item in the queue always describes the last state of the
CPU usage.

At RavenThreadPool- the ctor is created with DocumentDatabase in order to be able to track CPU usage changes per database

Added endpoints - DebugAutoTuningInfo in DebugController and also for AdminController.
Only Admin can get CPU usage statistics

Signed-off-by: Hila Mefano <hila@ayende.com>